+++ HISTORY.md
@@ -1,5 +1,10 @@
 # History
 
+## Version 0.1.1
+
+- Improvement: New class `Initializer` to facilitate the initialization proces
+- Improvement: New class `TradenityAppBundle`
+- Refactor: Move `SymfonySessionIdAccessor` to `Session` nameapace
 
 ## Version 0.1.0